Added Email authentication.
